requirements.txt
numpy==1.26.0
Flask==2.3.3
Werkzeug==2.3.8
line-bot-sdk==3.17.1
pandas==1.5.3
openpyxl==3.1.2
requests==2.32.3
python-dotenv==1.0.0
boto3==1.34.120
tensorflow-cpu==2.19.0
keras==3.10.0
Pillow